functions/create_subscription/pyproject.toml
[tool.poetry]
name = "create_subscription"
version = "0.1.0"
description = ""
authors = ["Joe Snell <joe.snell@daveramsey.com>"]

[tool.poetry.dependencies]
python = ">=3"
pyjwt = "^1.7.1"

[tool.poetry.dev-dependencies]
boto3 = "^1.13.18"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
